+++ src/lib.rs
@@ -1,7 +1,7 @@
-mod anchors;
 mod creds;
 mod encrypted_amounts;
 mod id_proofs;
+mod presentation;
 mod schema;
 mod transactions;
 mod types;
@@ -10,10 +10,10 @@ mod web3id;
 // UniFFI book: https://mozilla.github.io/uniffi-rs/udl_file_spec.html
 uniffi::include_scaffolding!("lib");
 
-pub use anchors::*;
 pub use creds::*;
 pub use encrypted_amounts::*;
 pub use id_proofs::*;
+pub use presentation::*;
 pub use schema::*;
 pub use transactions::*;
 pub use types::*;
